Update patch set 6
Patch Set 6:

> Jabeen - thanks for updating the user guide. However, the Java
 > Client user guide is now in the main documentation project. You and
 > Deven both were sent an email from me on 21 March ("Acumos Critical
 > Docs - Moving to Docs Repo") outlining the move - it's probably
 > buried in your inbox. I grabbed a copy of your updated guide from
 > gerrit. I am incorporating all the client guides into a single
 > onboarding guide. I submitted a patch on 21 March to remove the
 > user guide from the acumos-java-client repo: https://gerrit.acumos.org/r/#/c/1380/

Yes Aimee, Sorry if I am repeating. The user-guide has changed in this patch. So my question was why the -1? If everything is ok, we can merge. And you can take that new user-guide. Then we can make one last commit to remove the user-guide from this repo.

Patch-set: 6